fix(styles): split styles for better maintainability
