+++ frontend/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" href="/react-file-manager-logo.png" type="image/icon type" />
+    <link rel="icon" href="./public/react-file-manager-logo.png" type="image/icon type" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>React File Manager</title>
   </head>
