+++ frontend/src/App.scss
@@ -11,6 +11,6 @@ body {
 
   .file-manager-container {
     height: 70%;
-    width: 65%;
+    width: 67%;
   }
 }
