+++ frontend/src/File Manager/Bread Crumb/BreadCrumb.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react";
 import { MdHome, MdOutlineNavigateNext } from "react-icons/md";
+import "./BreadCrumb.scss";
 
 const BreadCrumb = ({ currentPath, setCurrentPath }) => {
   const [folders, setFolders] = useState([]);
@@ -21,11 +22,7 @@ const BreadCrumb = ({ currentPath, setCurrentPath }) => {
   return (
     <div className="breadcrumb">
       {folders.map((folder, index) => (
-        <span
-          key={index}
-          className="folder-name"
-          onClick={() => switchPath(index)}
-        >
+        <span key={index} className="folder-name" onClick={() => switchPath(index)}>
           {index === 0 ? (
             <>
               <MdHome /> Home
