+++ frontend/src/File Manager/Files/Files.scss
@@ -1,0 +1,263 @@
+@import "../../styles/variables";
+
+.files {
+  position: relative;
+  display: flex;
+  align-content: flex-start;
+  flex-wrap: wrap;
+  column-gap: 0.5em;
+  row-gap: 5px;
+  height: calc(94.2% - 16px);
+  @include overflow-y-scroll;
+  padding: 8px;
+  padding-right: 4px;
+
+  .file-item-container {
+    border-radius: 5px;
+  }
+
+  .file-selected {
+    background-color: rgb(0, 0, 0, 0.04);
+  }
+
+  .file-item {
+    position: relative;
+    height: 81px;
+    width: 138px;
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    align-items: center;
+    justify-content: space-between;
+    padding-top: 13px;
+    padding-bottom: 1px;
+    margin: 5px 0;
+    border-radius: 5px;
+
+    &:hover {
+      cursor: pointer;
+      background-color: rgb(0, 0, 0, 0.04);
+    }
+
+    .rename-file-container {
+      position: absolute;
+      top: 62px;
+      width: 100%;
+      text-align: center;
+    }
+
+    .rename-file-container.list {
+      top: 6px;
+      left: 48px;
+      text-align: left;
+
+      .rename-file {
+        min-width: 15%;
+        text-align: left;
+        border-radius: 3px;
+        border: none;
+        top: 5px;
+        white-space: nowrap;
+        overflow-x: hidden;
+        max-width: calc(100% - 48px);
+      }
+
+      .folder-name-error.right {
+        left: 0px;
+        bottom: -72px;
+      }
+    }
+
+    .rename-file {
+      position: relative;
+      min-width: 69%;
+      max-width: -webkit-fill-available;
+      text-align: center;
+      border: 4px solid $primary-color;
+      border-radius: 10px;
+      font-size: 15px;
+      font-family: "Nunito Sans", sans-serif;
+      padding: 2px 7px;
+      resize: none;
+      field-sizing: content;
+      overflow-y: hidden;
+
+      &:focus {
+        outline: none;
+      }
+    }
+
+    .folder-name-error {
+      position: absolute;
+      z-index: 1;
+      bottom: -68px;
+      left: 16px;
+      padding: 8px;
+      width: 292px;
+      border-radius: 5px;
+      background-color: #696969;
+      text-align: left;
+      margin: 0;
+      font-size: 0.9em;
+      &::before {
+        content: "";
+        position: absolute;
+        top: -20%;
+        rotate: -45deg;
+        border: 15px solid #696969;
+        border-color: transparent #696969 transparent transparent;
+      }
+    }
+
+    .folder-name-error.right {
+      left: 16px;
+      &::before {
+        left: 11%;
+      }
+    }
+
+    .folder-name-error.left {
+      left: -180px;
+      &::before {
+        left: 76%;
+        transform: rotate(90deg) scaleX(-1);
+      }
+    }
+
+    .folder-name-error.top {
+      bottom: unset !important;
+      top: -68px;
+      &::before {
+        content: none;
+      }
+      &:after {
+        content: "";
+        position: absolute;
+        bottom: -20%;
+        left: 11%;
+        rotate: -45deg;
+        border: 15px solid #696969;
+        border-color: transparent transparent #696969 transparent;
+      }
+    }
+
+    .folder-name-error.top.left {
+      &::after {
+        left: 76%;
+        transform: rotate(90deg) scaleX(-1);
+      }
+    }
+
+    .file-name {
+      max-width: 115px;
+    }
+  }
+
+  .file-selected {
+    background-color: $primary-color;
+    color: white;
+
+    &:hover {
+      cursor: pointer;
+      background-color: $primary-color;
+    }
+  }
+
+  .file-moving {
+    opacity: 0.5;
+  }
+}
+
+.files.list {
+  flex-direction: column;
+  flex-wrap: nowrap;
+  gap: 0;
+  padding-left: 0px;
+  padding-top: 0px;
+
+  .files-header {
+    font-size: 0.83em;
+    font-weight: 600;
+    display: flex;
+    gap: 5px;
+    border-bottom: 1px solid #dddddd;
+    padding: 4px 0;
+    position: sticky;
+    top: 0;
+    background-color: #f5f5f5;
+    z-index: 1;
+
+    .file-name {
+      width: calc(70% - 65px);
+      padding-left: 65px;
+    }
+
+    .file-date {
+      text-align: center;
+      width: 20%;
+    }
+
+    .file-size {
+      text-align: center;
+      width: 10%;
+    }
+  }
+
+  .file-item-container {
+    display: flex;
+    width: 100%;
+
+    &:hover {
+      cursor: pointer;
+      background-color: rgb(0, 0, 0, 0.04);
+    }
+  }
+
+  .file-item-container.file-selected {
+    &:hover {
+      background-color: $primary-color !important;
+    }
+  }
+
+  .file-item {
+    flex-direction: row;
+    height: 13px;
+    justify-content: unset;
+    padding: 15px;
+    padding-left: 20px;
+    margin: 0;
+    width: calc(70% - 30px);
+
+    &:hover {
+      background-color: unset;
+    }
+
+    .file-name {
+      max-width: 285px;
+    }
+  }
+
+  .modified-date {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 0.8em;
+    width: calc(20%);
+  }
+
+  .size {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 0.8em;
+    width: calc(10%);
+  }
+}
+
+.empty-folder {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+}
