+++ frontend/src/File Manager/Navigation Pane/NavigationPane.scss
@@ -1,0 +1,91 @@
+@import "../../styles/variables";
+
+.sb-folders-list {
+  list-style: none;
+  margin: 0px 4px;
+  height: 100%;
+  @include overflow-y-scroll;
+
+  .folder-collapsible {
+    margin-left: 10px;
+  }
+
+  .sb-folders-list-item {
+    display: flex;
+    align-items: center;
+    margin-bottom: 4px;
+    padding: 6px 5px;
+    border-radius: 4px;
+
+    &:hover {
+      cursor: pointer;
+      background-color: rgb(0, 0, 0, 0.04);
+    }
+
+    .non-expanable {
+      min-width: 20px;
+    }
+
+    .sb-folder-details {
+      display: flex;
+      align-items: center;
+
+      .folder-open-icon {
+        margin: 0 7px;
+      }
+
+      .folder-close-icon {
+        margin: 1px 9px 0px 8px;
+      }
+
+      .sb-folder-name {
+        width: max-content;
+      }
+    }
+
+    .folder-rotate-down {
+      animation: rotateFolderDown 0.5s linear forwards;
+    }
+
+    @keyframes rotateFolderDown {
+      from {
+        transform: rotate(0deg);
+      }
+
+      to {
+        transform: rotate(90deg);
+      }
+    }
+
+    .folder-rotate-right {
+      animation: rotateRight 0.5s linear forwards;
+    }
+
+    @keyframes rotateRight {
+      from {
+        transform: rotate(90deg);
+      }
+
+      to {
+        transform: rotate(0deg);
+      }
+    }
+  }
+
+  .active-list-item {
+    background-color: $primary-color;
+    color: white;
+
+    &:hover {
+      cursor: pointer;
+      background-color: $primary-color;
+    }
+  }
+
+  .empty-nav-pane {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+  }
+}
